backend/requirements.txt
# Web framework
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1

# Data parsing and scraping
beautifulsoup4==4.12.2
requests==2.31.0
scrapy==2.11.0
selenium==4.15.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Task scheduling
apscheduler==3.10.4

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.2

# Data processing
pandas==2.1.4
numpy==1.26.2

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-mock==3.12.0

# Code quality
flake8==6.1.0
black==23.12.1

# Logging
loguru==0.7.2